% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bulk_email_checker.R
\name{bulk_email_checker}
\alias{bulk_email_checker}
\title{bulk_email_checker}
\usage{
bulk_email_checker(email, key, sleep = 0)
}
\arguments{
\item{email}{\dots}

\item{key}{\dots}

\item{sleep}{\dots}
}
\description{
bulk_email_checker
}
