% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email_marker.R
\name{email_marker}
\alias{email_marker}
\title{email_marker}
\usage{
email_marker(email, key, sleep = 0)
}
\arguments{
\item{email}{\dots}

\item{key}{\dots}

\item{sleep}{\dots}
}
\description{
email_marker
}
