% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_flow.R
\name{listflow_post_contacts_file}
\alias{listflow_post_contacts_file}
\title{listflow_post_contacts_file}
\usage{
listflow_post_contacts_file(file, csv_file)
}
\arguments{
\item{file}{\dots}

\item{csv_file}{\dots}
}
\description{
listflow_post_contacts_file
}
