% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_email_verification.R
\name{quick_email_verification}
\alias{quick_email_verification}
\title{quick_email_verification}
\usage{
quick_email_verification(email, key, sleep = 0)
}
\arguments{
\item{email}{\dots}

\item{key}{\dots}

\item{sleep}{\dots}
}
\description{
quick_email_verification
}
