% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_hit.R
\name{spot_hit_add_short_url}
\alias{spot_hit_add_short_url}
\title{spot_hit_short_url}
\usage{
spot_hit_add_short_url(
  adresse_cible,
  extension,
  key,
  nom = NULL,
  domaine_id = NULL,
  sous_domaine = NULL
)
}
\arguments{
\item{adresse_cible}{\dots}

\item{extension}{\dots}

\item{key}{API key}

\item{nom}{\dots}

\item{domaine_id}{\dots}

\item{sous_domaine}{\dots}
}
\description{
spot_hit_short_url
}
