% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_hit.R
\name{spot_hit_remove_short_url}
\alias{spot_hit_remove_short_url}
\title{spot_hit_remove_short_url}
\usage{
spot_hit_remove_short_url(id, key)
}
\arguments{
\item{id}{\dots}

\item{api_key}{API key}
}
\description{
spot_hit_remove_short_url
}
