% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_hit.R
\name{spot_hit_send_sms}
\alias{spot_hit_send_sms}
\title{spot_hit_send_sms}
\usage{
spot_hit_send_sms(
  message,
  destinataires,
  key,
  expediteur = NULL,
  type = "premium",
  smslong = 0,
  encodage = "auto",
  nom = NULL
)
}
\arguments{
\item{message}{\dots}

\item{destinataires}{\dots}

\item{key}{API key}

\item{expediteur}{\dots}

\item{type}{premium or lowcost}

\item{smslong}{\dots}

\item{encodage}{\dots}

\item{nom}{\dots}
}
\description{
spot_hit_send_sms
}
